+++ CHANGELOG.md
@@ -6,11 +6,15 @@
 
 - Added documentation for image utilities.
 - Added image utility .img-thumbnail.
+- Added size classes for material icons.
 
 ### Changed
 
+- Renamed icons to iconography under core, and moved payment icons under material icons.
 - Slightly increased the bottom margin on all form fields to allow for better grouping and spacing for data attribute messages.
 
 ### Fixed
 
 ### Removed
+
+- Removed Payment Icons from core documentation.